refactor: move stack functions to subfolder
