+++ src/stack/compute/__tests__/__image_snapshots__/median-image-test-ts-more-complex-stack-1-snap.png
