+++ src/stack/compute/__tests__/histogram.test.ts
@@ -1,4 +1,4 @@
-import { Stack } from '../../Stack';
+import { Stack } from '../../../Stack';
 
 test('two grey images, bitsDepth = 8', () => {
   const image = testUtils.createGreyImage([[1, 2, 3, 4]]);
