+++ src/stack/compute/__tests__/medianImage.test.ts
@@ -1,8 +1,8 @@
 import { join } from 'node:path';
 
-import { Image } from '../../Image';
-import { Stack } from '../../Stack';
-import { getStackFromFolder } from '../utils/getStackFromFolder';
+import { Image } from '../../../Image';
+import { Stack } from '../../../Stack';
+import { getStackFromFolder } from '../../utils/getStackFromFolder';
 
 test('3 grey images', () => {
   const image1 = testUtils.createGreyImage([[1, 2, 3, 4]]);
