+++ src/stack/compute/index.ts
@@ -1,4 +1,6 @@
+export * from './histogram';
 export * from './maxImage';
-export * from './minImage';
 export * from './meanImage';
 export * from './medianImage';
+export * from './minImage';
+export * from './sum';
